Ingest From Message Working
This sets up MinIO + Console in order to be able to use MinIO
correctly. This sets up a basic Python worker that is able to
pull the image from the URL, upload that to MinIO, and supposedly
MinIO will then push to an AMQP queue for some other system to
listen to.

That next system will be responsible for running Docling on the
documents and doing the needful there.